prevent duplicate aliases (settings banch)
Added a check to the autoAliases method to prevent aliases from being added if it is already present in the aliases array. Possibly added by the user manually or if you have aliases like `this-is-an-example`, `this-isan-example`, `thisisan-example` which would normally end up changing them all to `thisisanexample` and adding them to the aliases array.