remove private message#levelID
the command in the inhibitor/finalizer is not necessarily the message.command if for instance it is a reaction command. Therefore, that command has to be taken account for, not the one potentually on message, making levelID useless. Since it's private nobody should be using it, therefore it's not breaking to remove it.